Split GMS in HMS integrations in separate modules
