+++ build.gradle
@@ -9,7 +9,7 @@ buildscript {
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:4.2.1'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.5.10"
+        classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:1.5.31'
         classpath "de.mannodermaus.gradle.plugins:android-junit5:1.7.1.1"
 
         classpath 'com.vanniktech:gradle-maven-publish-plugin:0.14.2'
