+++ core/proguard-rules-consumer.pro
