+++ core/src/main/AndroidManifest.xml
