+++ core/src/main/java/com/patloew/colocation/TaskCancelledException.kt
