--- library/build.gradle
@@ -1,68 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-    id 'de.mannodermaus.android-junit5'
-    id 'jacoco'
-}
-
-ext {
-    RELEASE_REPOSITORY_URL = "https://s01.oss.sonatype.org/service/local/staging/deploy/maven2/"
-    SNAPSHOT_REPOSITORY_URL = "https://s01.oss.sonatype.org/content/repositories/snapshots/"
-}
-
-apply plugin: "com.vanniktech.maven.publish"
-
-jacoco {
-    toolVersion = "0.8.7"
-}
-
-android {
-    compileSdkVersion 31
-
-    defaultConfig {
-        minSdkVersion 19
-        targetSdkVersion 31
-        versionCode 1
-        versionName "1.1.0"
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-            consumerProguardFiles 'proguard-rules-consumer.pro'
-        }
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
-    testOptions {
-        unitTests.returnDefaultValues = true
-    }
-    lintOptions {
-        abortOnError false
-    }
-}
-
-dependencies {
-    api fileTree(dir: 'libs', include: ['*.jar'])
-
-    api 'com.google.android.gms:play-services-location:18.0.0'
-    api 'com.huawei.hms:location:6.0.0.302'
-
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:1.5.10"
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.5.0"
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-play-services:1.5.0"
-
-    testImplementation "io.mockk:mockk:1.11.0"
-    testImplementation "org.junit.jupiter:junit-jupiter-api:5.7.2"
-    testRuntimeOnly "org.junit.jupiter:junit-jupiter-engine:5.7.2"
-    testImplementation "org.junit.jupiter:junit-jupiter-params:5.7.2"
-    testImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.5.0'
-
-}
-
-task jacocoTestReport(type: JacocoReport, dependsOn: ['testDebugUnitTest'])
