+++ sample/build.gradle
@@ -55,8 +55,8 @@ dependencies {
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.5.0"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.5.0"
 
-    implementation project(':library')
-    //implementation 'com.patloew.colocation:colocation:1.1.0'
+    implementation project(':google')
+    implementation project(':huawei')
 
     implementation 'com.google.android.gms:play-services-location:18.0.0'
