+++ sample/src/main/res/layout/activity_main.xml
@@ -54,31 +54,4 @@
         android:text="Location source"
         style="@style/TextAppearance.AppCompat.Caption"/>
 
-    <RadioGroup
-        android:id="@+id/source_group"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-        <RadioButton
-            android:id="@+id/source_auto"
-            android:text="Choose automatically"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:checked="true"/>
-
-        <RadioButton
-            android:id="@+id/source_gms"
-            android:text="GMS"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:checked="false"/>
-
-        <RadioButton
-            android:id="@+id/source_hms"
-            android:text="HMS"
-            android:checked="false"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"/>
-    </RadioGroup>
-
-
 </LinearLayout>
